feat: finish builtins

+++ srcs/exec/glob_exp.c
@@ -6,7 +6,7 @@
 /*   By: rgomes-d <rgomes-d@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/05 14:16:04 by rgomes-d          #+#    #+#             */
-/*   Updated: 2026/01/08 14:54:55 by rgomes-d         ###   ########.fr       */
+/*   Updated: 2026/01/10 14:48:09 by rgomes-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,7 +78,7 @@ int	check_literal_match(t_segment *aux_seg, char *content, unsigned long i)
 		len_s2 = ft_strlen(aux_seg->text);
 		if (!aux_seg->next && (len_s1 - len_s2) >= (int)i)
 			i = len_s1 - len_s2;
-		if (i >= 0 && !ft_strncmp(&content[i], aux_seg->text, len_s2))
+		if ((long)i >= 0 && !ft_strncmp(&content[i], aux_seg->text, len_s2))
 			return (1);
 	}
 	return (0);
